% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_encode}
\alias{base64_decode}
\alias{base64_encode}
\alias{base64url_decode}
\alias{base64url_encode}
\title{Encode and decode base64}
\usage{
base64_encode(bin, linebreaks = FALSE)

base64_decode(text)

base64url_encode(bin)

base64url_decode(text)
}
\arguments{
\item{bin}{raw or character vector with data to encode into base64}

\item{linebreaks}{insert linebreaks in the base64 message to make it more readable}

\item{text}{string with base64 data to decode}
}
\description{
Encode and decode binary data into a base64 string. Character vectors are
automatically collapsed into a single string.
}
\details{
The \code{base64url_encode} functions are specified in Section 5 of RFC 4648
as a URL-safe alternative to the standard base64. They use different characters
for the 62:nd and 63:rd alphabet character and don't include trailing \code{==}
padding.
}
\examples{
input <- charToRaw("foo = bar + 5")
message <- base64_encode(input)
output <- base64_decode(message)
identical(output, input)
}

