% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x25519.R
\name{curve25519}
\alias{curve25519}
\alias{read_ed25519_key}
\alias{read_ed25519_pubkey}
\alias{read_x25519_key}
\alias{read_x25519_pubkey}
\alias{ed25519_sign}
\alias{ed25519_verify}
\alias{x25519_diffie_hellman}
\title{Curve25519}
\usage{
read_ed25519_key(x)

read_ed25519_pubkey(x)

read_x25519_key(x)

read_x25519_pubkey(x)

ed25519_sign(data, key)

ed25519_verify(data, sig, pubkey)

x25519_diffie_hellman(key, pubkey)
}
\arguments{
\item{x}{a 32 byte raw vector with (pub)key data}

\item{key}{private key as returned by [read_ed25519_key] or [ed25519_keygen]}

\item{pubkey}{public key as returned by [read_ed25519_pubkey] or `key$pubkey`}
}
\description{
Curve25519 is a recently added low-level algorithm that can be used both for
diffie-hellman (called X25519) and for signatures (called ED25519). Note that
these functions are only available when building against version 1.1.1 or
newer of the openssl library. The same functions are also available in the
sodium R package.
}
