% Generated by roxygen2 (4.0.2): do not edit by hand
\name{digest}
\alias{digest}
\alias{digest_string}
\title{Digest a string}
\usage{
digest(x, algo)

digest_string(x, algo, string = FALSE)
}
\arguments{
\item{x}{strings to digest}

\item{algo}{digest algorithm, e.g. 'md5' 'sha1', 'sha256', etc.}

\item{string}{converts output value into a string (TRUE/FALSE)}
}
\description{
The \code{digest_string} function creates a raw hash from a single string.
The \code{digest} function is a vectorised version that creates digest strings
from many strings at once.
}
\examples{
openssl::digest("foo", "md5")
digest::digest("foo", "md5", serialize = FALSE)
}
\references{
OpenSSL manual: \url{https://www.openssl.org/docs/crypto/EVP_DigestInit.html}
}

