% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssl.R
\name{download_cert}
\alias{download_cert}
\title{Download Certificate}
\usage{
download_cert(host = "localhost", port = 443)
}
\arguments{
\item{host}{string: hostname of the server to connect to}

\item{port}{integer: port to connect to}
}
\description{
Connects to a HTTPS server and retrieves the SSL certificate.
}

