% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecc.R
\name{ec}
\alias{ec}
\alias{ec_curve}
\alias{ec_secp256k1}
\title{Eliptic Curves}
\usage{
ec_curve(a, b, p)

ec_secp256k1()
}
\arguments{
\item{a}{bignum a}

\item{b}{bignum b}

\item{p}{bignum p}
}
\description{
Bla bla stuff
}

