% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmac.R
\name{hmac}
\alias{hmac}
\title{HMAC}
\usage{
hmac(data, key, hash = c("md5", "sha1", "sha256", "sha512"))
}
\arguments{
\item{data}{raw vector with data to sign}

\item{key}{raw vector with secret key}

\item{hash}{which digest function to use}
}
\description{
HMAC is a MAC (message authentication code), i.e. a keyed hash function.
}
\examples{
key <- rand_bytes(16)
msg <- serialize(iris, NULL)
hmac(msg, key, "md5")
hmac(msg, key, "sha1")
}
\references{
\url{https://www.openssl.org/docs/manmaster/crypto/hmac.html}
}

