% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keygen.R
\name{keygen_rsa}
\alias{keygen_dsa}
\alias{keygen_ecdsa}
\alias{keygen_rsa}
\title{Generate Key pair}
\usage{
keygen_rsa(bits = 2048)

keygen_dsa(bits = 2048)

keygen_ecdsa(curve = c("P-256", "P-384", "P-521"))
}
\arguments{
\item{bits}{bitsize of the generated key}

\item{curve}{which NIST curve to use}
}
\description{
The \code{keygen} functions generate a random private key. Use \code{as.list(key)$pubkey}
to derive the corresponding public key.
}
\examples{
key <- keygen_rsa(1024)
pubkey <- as.list(key)$pubkey
}

