% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykey.R
\name{my_key}
\alias{my_key}
\alias{my_pubkey}
\title{Default keypair}
\usage{
my_key()

my_pubkey()
}
\description{
The \code{my_key()} function looks for environment variable \code{USER_KEY} with a
path to a private keyfile. If unset it defaults to \code{"~/.ssh/id_rsa"}. The
\code{my_pubkey()} function looks for environment variable \code{USER_PUBKEY} which
defaults to \code{"~/.ssh/id_rsa.pub"}. If neither exists it will also try \code{my_key()}
and derive the corresponding public key.
}

