% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{openssl}
\alias{openssl}
\alias{openssl-package}
\title{R bindings for the OpenSSL library}
\description{
Interfaces to the OpenSSL libraries libssl and libcrypto.
Currently only implements bindings to the OpenSSL secure random number
generator and digest functions.
}
\author{
Jeroen Ooms, Oliver Keyes
}

