% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsa.R
\name{parse_key}
\alias{parse_key}
\title{Parse RSA keys}
\usage{
parse_key(text, password = readlines("Enter password"))
}
\arguments{
\item{text}{a path or text string containing}
}
\description{
Parse a text string containing a public or private RSA key in the
typical base64 encoded format.
}

