% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcrypt.R
\name{bcrypt_pbkdf}
\alias{bcrypt_pbkdf}
\title{Bcrypt PWKDF}
\usage{
bcrypt_pbkdf(password, salt, rounds = 16L, size = 32L)
}
\arguments{
\item{password}{string or raw vector with password}

\item{salt}{raw vector with (usually 16) bytes}

\item{rounds}{number of hashing rounds}

\item{size}{desired length of the output key}
}
\description{
Password based key derivation function with bcrypt. This is not
part of openssl. It is needed to parse private key files which are
encoded in the \href{https://cvsweb.openbsd.org/src/usr.bin/ssh/PROTOCOL.key?annotate=HEAD}{new openssh format}.
}
