% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pem.R
\name{read_der}
\alias{read_der}
\alias{read_pem}
\alias{write_pem}
\title{Read PEM/DER files}
\usage{
read_der(file, type = c("guess", "cert", "key", "pubkey"))

read_pem(file, multiple = FALSE, password = readline)

write_pem(bin)
}
\arguments{
\item{file}{a connection, file path or vector with literal data}

\item{multiple}{read multiple PEM keys or certificates from a single file}

\item{password}{a string or callback function}

\item{bin}{binary DER representation of a key or cert}
}
\description{
DER is a binary format used for storing certificate and key data in SSL.
The PEM format is base64 encoded DER data surrounded by header lines and
possibly password protected. It is the most commonly used format to exchange
RSA keys and X509 certificates. It is easily recognized from the
\code{-- BEGIN --} and \code{-- END --} lines.
}
\details{
A PEM file can contain multiple values, for example to store a certificate
chain. These functions parse and validates the PEM or DER input and return
a raw vector which holds the DER data.
}

