% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkcs7.R
\name{pkcs7_encrypt}
\alias{pkcs7_encrypt}
\alias{pkcs7_decrypt}
\title{Encrypt/decrypt pkcs7 messages}
\usage{
pkcs7_encrypt(message, cert, pem = TRUE)

pkcs7_decrypt(input, key, der = is.raw(input))
}
\arguments{
\item{message}{text or raw vector with data to encrypt}

\item{cert}{the certificate with public key to use for encryption}

\item{pem}{convert output pkcs7 data to PEM format}

\item{input}{file path or string with PEM or raw vector with p7b data}

\item{key}{private key to decrypt data}

\item{der}{assume input is in DER format (rather than PEM)}
}
\description{
Encrypt or decrypt messages using PKCS7 smime format.
Note PKCS7 only supports RSA keys.
}
\seealso{
\link{encrypt_envelope}
}
