% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rand_num}
\alias{rand_num}
\title{Generate random numbers}
\usage{
rand_num(n)
}
\arguments{
\item{n}{how many random numbers to generate}
}
\description{
Simple wrapper for rand_bytes to generate random 32 bit doubles
between 0 and 1. This is a somewhat more convenient and conventional
interface to the RNG.
}
\examples{
# Use CDF to map [0,1] into random draws from a distribution
x <- qnorm(rand_num(1000), mean=100, sd=15)
hist(x)

y <- qbinom(rand_num(1000), size=10, prob=0.3)
hist(y)
}

