% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pem.R
\name{read_pem}
\alias{read_pem}
\title{Read PEM files}
\usage{
read_pem(file, password = readline)
}
\arguments{
\item{file}{a connection, file path or character vector with PEM data}

\item{password}{a string or callback function}
}
\description{
The PEM format is simply base64 encoded data surrounded by header lines.
It is the most commonly used format to store RSA keys and X509 certificates.
It is easily recognized from the \code{-- BEGIN --} and \code{-- END --}
lines.
}
\details{
This function parses and validates the PEM file and returns the binary DER
representation.
}

