% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sslctx.R
\name{ssl_ctx_info}
\alias{ssl_ctx_info}
\title{Reading an SSL_CTX pointer}
\usage{
ssl_ctx_info(ptr)
}
\arguments{
\item{ptr}{an external pointer object of class ssl_ctx}
}
\description{
Reads data from a pointer to an SSL_CTX object that is provided
via another library such as curl.
}
