% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_pem}
\alias{write_der}
\alias{write_pem}
\title{Export key or certificate}
\usage{
write_pem(x, path = NULL, password = readline)

write_der(x, path = NULL)
}
\arguments{
\item{x}{a public/private key or certificate object}

\item{path}{file to write to. If \code{NULL} it returns the output as a string.}

\item{password}{string or callback function to set password (only applicable for
private keys).}
}
\description{
The \code{write_pem} functions exports a key or certificate to the standard
base64 PEM format. For private keys it is possible to set a password.
}

