% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64.R
\name{base64_encode}
\alias{base64_encode}
\alias{base64_decode}
\title{Encode and decode base64}
\usage{
base64_encode(bin, linebreaks = FALSE)

base64_decode(text)
}
\arguments{
\item{bin}{raw or character vector with data to encode into base64}

\item{linebreaks}{insert linebreaks in the base64 message to make it more readable}

\item{text}{string with base64 data to decode}
}
\description{
Encode and decode binary data into a base64 string. Character vectors are
automatically collapsed into a single string.
}
\examples{
input <- charToRaw("foo = bar + 5")
message <- base64_encode(input)
output <- base64_decode(message)
identical(output, input)
}
