% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cert.R
\name{cert_verify}
\alias{ca_bundle}
\alias{cert_verify}
\alias{ssl_cert_chain}
\title{X509 certificates}
\usage{
cert_verify(cert, root = ca_bundle())

ssl_cert_chain(host = "localhost", port = 443)

ca_bundle()
}
\arguments{
\item{cert}{certficate (or certificate-chain) to be verified. Must be cert or list or path.}

\item{root}{trusted pubkey or certificate(s) e.g. CA bundle.}

\item{host}{string: hostname of the server to connect to}

\item{port}{integer: port to connect to}
}
\description{
Read, download and verify certificates.
}
\examples{
# Verify the r-project HTTPS cert
cert <- ssl_cert_chain("www.r-project.org", 443)
print(cert)
print(as.list(cert[[1]])$pubkey)
cert_verify(cert, ca_bundle())
}

