% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cert.R
\name{cert_verify}
\alias{ca_bundle}
\alias{cert_verify}
\title{Certificates}
\usage{
cert_verify(chain, root = ca_bundle())

ca_bundle()
}
\arguments{
\item{chain}{certficate (or certificate-chain) to be verified. Must be cert or list or path.}

\item{root}{trusted root certificates (e.g. CA bundle). Must be cert or list or path.}
}
\description{
Verify a certificate
}

