% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cert.R
\name{certinfo}
\alias{certinfo}
\alias{certs}
\alias{verify_cert}
\title{Certificates}
\usage{
certinfo(cert)

verify_cert(cert, ca)
}
\arguments{
\item{cert}{a certificate}

\item{ca}{a CA bundle}
}
\description{
Stuff for certificates
}

