% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hash}
\alias{hash}
\alias{rawhash}
\title{Hash a string}
\usage{
hash(x, algo)

rawhash(x, algo)
}
\arguments{
\item{x}{strings to digest}

\item{algo}{digest algorithm, e.g. 'md5' 'sha1', 'sha256', etc.}
}
\description{
The \code{rawhash} function creates a raw hash from a single string.
The \code{hash} function is a vectorised version that creates digest strings
from many strings at once.
}
\examples{
library(digest)
hash("foo", "md5")
digest("foo", "md5", serialize = FALSE)
}
\references{
OpenSSL manual: \url{https://www.openssl.org/docs/crypto/EVP_DigestInit.html}
}

