% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keygen.R
\name{keygen}
\alias{dsa_keygen}
\alias{ec_keygen}
\alias{keygen}
\alias{rsa_keygen}
\title{Generate Key pair}
\usage{
rsa_keygen(bits = 2048)

dsa_keygen(bits = 1024)

ec_keygen(curve = c("P-256", "P-384", "P-521"))
}
\arguments{
\item{bits}{bitsize of the generated RSA/DSA key}

\item{curve}{which NIST curve to use}
}
\description{
The \code{keygen} functions generate a random private key. Use \code{as.list(key)$pubkey}
to derive the corresponding public key.
}
\examples{
key <- rsa_keygen()
pubkey <- as.list(key)$pubkey
}

