% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mykey.R
\name{my_key}
\alias{my_key}
\alias{my_pubkey}
\title{Default keypair}
\usage{
my_key()

my_pubkey()
}
\description{
The default key can be configured using environment variable \code{USER_KEY}
which points to the user (private) key. If unset it defaults to \code{"~/.ssh/id_rsa"}.
The \code{my_pubkey} function derives the corresponding public key.
}

