% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openssl.R
\docType{package}
\name{openssl}
\alias{openssl}
\alias{openssl-package}
\title{R bindings for the OpenSSL library}
\description{
Interfaces to the OpenSSL libraries libssl and libcrypto.
Currently implements bindings to the OpenSSL secure random number
generator and cryptographic hash functions.
}
\author{
Jeroen Ooms, Oliver Keyes
}

