% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_key}
\alias{read_cert}
\alias{read_key}
\alias{read_pubkey}
\title{Read keys and certificates}
\usage{
read_key(file, password = readline, format = c("pem", "der"))

read_pubkey(file, password = readline, format = c("pem", "der", "ssh"))

read_cert(file, format = c("pem", "der"))
}
\description{
Read from a file or buffer in DER or PEM format.
}

