% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signing.R
\name{md5_sign}
\alias{md5_sign}
\alias{md5_verify}
\alias{sha1_sign}
\alias{sha1_verify}
\alias{sha256_sign}
\alias{sha256_verify}
\title{Signatures}
\usage{
md5_sign(hash, key = my_key(), password = readline)

sha1_sign(hash, key = my_key(), password = readline)

sha256_sign(hash, key = my_key(), password = readline)

md5_verify(hash, sig, pubkey = my_pubkey())

sha1_verify(hash, sig, pubkey = my_pubkey())

sha256_verify(hash, sig, pubkey = my_pubkey())
}
\arguments{
\item{hash}{string or raw vector with md5, sha1 or sha256 hash}

\item{key}{private key or file path. See \code{\link{read_key}}.}

\item{password}{string or a function to read protected keys.}

\item{sig}{path or raw vector with signature data}

\item{pubkey}{public key or file path. See \code{\link{read_pubkey}}.}
}
\description{
Sign and verify a message digest. DSA and ECDSA keys only support SHA1
and SHA256 signatures.
}
\examples{
\dontrun{
hash <- sha256(system.file("DESCRIPTION"))
sig <- sha256_sign(hash)
sha256_verify(hash, sig)

hash <- sha1(serialize(iris, NULL))
sig <- sha1_sign(hash)
sha1_verify(hash, sig)
}
}

