% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signing.R
\name{signing}
\alias{message_sign}
\alias{message_verify}
\alias{rsa_sign}
\alias{rsa_verify}
\alias{signing}
\title{Sign and verify}
\usage{
rsa_sign(hash, key = "~/.ssh/id_rsa", password = readline)

rsa_verify(hash, sig, pubkey = "~/.ssh/id_rsa.pub")

message_sign(msg, key, password = readline)

message_verify(msg, sig, pubkey)
}
\arguments{
\item{hash}{string or raw vector with md5, sha1 or sha256 hash}

\item{key}{file path or raw/character vector with RSA private key}

\item{password}{either a hardcoded string or a callback function that
returns a string. Only required if key is protected with a passphrase.}

\item{sig}{raw vector with signature data}

\item{pubkey}{file path or raw/character vector with RSA public key}
}
\description{
Create and verify RSA signatures.
}
\examples{
\dontrun{
hash <- sha256(system.file("DESCRIPTION"))
sig <- rsa_sign(hash)
rsa_verify(hash, sig)

hash <- sha1(serialize(iris, NULL))
sig <- rsa_sign(hash)
rsa_verify(hash, sig)

hash <- md5("i like cookies")
sig <- rsa_sign(hash)
rsa_verify(hash, sig)
}
}

